feat: :sparkles: Pointers
Pointers in Go are powerful tools that allow you to manipulate variables and data structures directly in memory. By storing the memory address of a variable, you can pass the address as a pointer to a function, modify the value of the variable, and have those changes reflected outside of the function. This makes Go an efficient language for handling large amounts of data and allows for flexible and dynamic programming.