feat: :sparkles: Clousures
Closures in Go are a powerful feature that allow you to capture and manipulate external variables within a function. By creating a closure, you can store and maintain the state of a function and access it across multiple calls. This allows for flexible and dynamic programming, and closures are commonly used to implement iterators, handlers, and callbacks